refactor: Make `SelectMultiple` component
