+++ editor.planx.uk/src/@planx/components/FileUploadAndLabel/SelectMultipleFileTypes.tsx
@@ -1,22 +1,15 @@
-import ArrowIcon from "@mui/icons-material/KeyboardArrowDown";
-import Autocomplete, {
+import {
   AutocompleteChangeReason,
-  autocompleteClasses,
   AutocompleteProps,
 } from "@mui/material/Autocomplete";
 import Box from "@mui/material/Box";
-import FormControl from "@mui/material/FormControl";
-import { inputLabelClasses } from "@mui/material/InputLabel";
 import List from "@mui/material/List";
 import ListItem from "@mui/material/ListItem";
 import ListSubheader from "@mui/material/ListSubheader";
-import { outlinedInputClasses } from "@mui/material/OutlinedInput";
-import { styled } from "@mui/material/styles";
-import TextField from "@mui/material/TextField";
 import Typography from "@mui/material/Typography";
 import capitalize from "lodash/capitalize";
 import React, { forwardRef, PropsWithChildren, useMemo } from "react";
-import { borderedFocusStyle } from "theme";
+import { CustomCheckbox, SelectMultiple } from "ui/shared/SelectMultiple";
 
 import { FileUploadSlot } from "../FileUpload/model";
 import {
@@ -36,100 +29,6 @@ interface SelectMultipleProps {
 interface Option extends UserFile {
   category: keyof FileList;
 }
-
-const StyledAutocomplete = styled(
-  Autocomplete<Option, true, true, false, "div">,
-)(({ theme }) => ({
-  marginTop: theme.spacing(2),
-  // Prevent label from overlapping expand icon
-  "& > div > label": {
-    paddingRight: theme.spacing(3),
-  },
-  // Vertically center "large" size caret icon
-  [`& .${autocompleteClasses.endAdornment}`]: {
-    top: "unset",
-  },
-  "&:focus-within": {
-    "& svg": {
-      color: "black",
-    },
-  },
-}));
-
-const StyledTextField = styled(TextField)(({ theme }) => ({
-  "&:focus-within": {
-    ...borderedFocusStyle,
-    [`& .${outlinedInputClasses.notchedOutline}`]: {
-      border: "1px solid transparent !important",
-    },
-  },
-  [`& .${outlinedInputClasses.notchedOutline}`]: {
-    borderRadius: 0,
-    border: `1px solid${theme.palette.border.main} !important`,
-  },
-  "& fieldset": {
-    borderColor: theme.palette.border.main,
-  },
-  backgroundColor: theme.palette.background.paper,
-  [`& .${outlinedInputClasses.root}, input`]: {
-    cursor: "pointer",
-  },
-  [`& .${inputLabelClasses.root}`]: {
-    textDecoration: "underline",
-    color: theme.palette.primary.main,
-    "&[data-shrink=true]": {
-      textDecoration: "none",
-      color: theme.palette.text.primary,
-      paddingY: 0,
-      transform: "translate(0px, -22px) scale(0.85)",
-    },
-  },
-}));
-
-const CustomCheckbox = styled("span")(({ theme }) => ({
-  display: "inline-flex",
-  flexShrink: 0,
-  position: "relative",
-  width: 40,
-  height: 40,
-  borderColor: theme.palette.text.primary,
-  border: "2px solid",
-  background: "transparent",
-  marginRight: theme.spacing(1.5),
-  "&.selected::after": {
-    content: "''",
-    position: "absolute",
-    height: 24,
-    width: 12,
-    borderColor: theme.palette.text.primary,
-    borderBottom: "5px solid",
-    borderRight: "5px solid",
-    left: "50%",
-    top: "42%",
-    transform: "translate(-50%, -50%) rotate(45deg)",
-    cursor: "pointer",
-  },
-}));
-
-/**
- * Function which returns the Input component used by Autocomplete
- */
-const renderInput: AutocompleteProps<
-  Option,
-  true,
-  true,
-  false,
-  "div"
->["renderInput"] = (params) => (
-  <StyledTextField
-    {...params}
-    InputProps={{
-      ...params.InputProps,
-      notched: false,
-    }}
-    label="What does this file show? (select all that apply)"
-  />
-);
 
 /**
  * Function which returns the groups (ul elements) used by Autocomplete
@@ -180,8 +79,6 @@ const renderOption: AutocompleteProps<
   </ListItem>
 );
 
-const PopupIcon = <ArrowIcon sx={{ color: "primary.main" }} fontSize="large" />;
-
 /**
  * Custom Listbox component
  * Used to wrap options within the autocomplete and append a custom element above the option list
@@ -224,7 +121,7 @@ export const SelectMultipleFileTypes = (props: SelectMultipleProps) => {
   const value: Option[] = useMemo(
     () =>
       initialTags.flatMap((tag) => options.filter(({ name }) => name === tag)),
-    [initialTags],
+    [initialTags, options],
   );
 
   /**
@@ -266,46 +163,19 @@ export const SelectMultipleFileTypes = (props: SelectMultipleProps) => {
   };
 
   return (
-    <FormControl
+    <SelectMultiple<Option>
+      getOptionLabel={(option) => option.name}
+      groupBy={(option) => option.category}
+      id={`select-multiple-file-tags-${uploadedFile.id}`}
+      isOptionEqualToValue={(option, value) => option.name === value.name}
       key={`form-${uploadedFile.id}`}
-      sx={{ display: "flex", flexDirection: "column" }}
-    >
-      <StyledAutocomplete
-        role="status"
-        aria-atomic={true}
-        aria-live="polite"
-        disableClearable
-        disableCloseOnSelect
-        getOptionLabel={(option) => option.name}
-        groupBy={(option) => option.category}
-        id={`select-multiple-file-tags-${uploadedFile.id}`}
-        isOptionEqualToValue={(option, value) => option.name === value.name}
-        ListboxComponent={ListboxComponent}
-        multiple
-        onChange={handleChange}
-        options={options}
-        popupIcon={PopupIcon}
-        renderGroup={renderGroup}
-        renderInput={renderInput}
-        renderOption={renderOption}
-        value={value}
-        ChipProps={{
-          variant: "uploadedFileTag",
-          size: "small",
-          sx: { pointerEvents: "none" },
-          onDelete: undefined,
-        }}
-        componentsProps={{
-          popupIndicator: {
-            disableRipple: true,
-          },
-          popper: {
-            sx: {
-              boxShadow: 10,
-            },
-          },
-        }}
-      />
-    </FormControl>
+      label="What does this file show? (select all that apply)"
+      ListboxComponent={ListboxComponent}
+      onChange={handleChange}
+      options={options}
+      renderGroup={renderGroup}
+      renderOption={renderOption}
+      value={value}
+    />
   );
 };
